MXS-3401: Only use servers parameter during startup
The parameter values should currently only be used when the service is
starting up and is being created from a configuration file. At runtime,
the targets are added individually and aren't done via the configuration.

Moved the post-configuration work into the correct function. This makes
sure it's always done whenever the configuration is updated.

Fixed the JSON parameter generation to remove the object type parameters
(servers etc.) and to mask the password in the retuned object.