Always assign an ID for COM_STMT_PREPARE
Now that the backends rely on the client protocol to direct the prepared
statement ID generation, an ID must be generated even if no history is
being recorded. This allows routers to use prepared statements even if
they don't support the collection of session command history.

Also adjusted the debug assertion for unknown prepared statements to
expect the case where a COM_STMT_CLOSE is sent with an ID of 0. This
happens when the ps_execute_direct test is run and is expected behavior.