Reorder REST API path checks for modules
Whenever a :module path segment was being resolved, the first thing it
attempted to do was to load the module. This would result in errors if the
module was one of the built-in core modules (i.e. maxscale or
servers). Doing the module loading last prevents these errors from being
generated.