MXS-3401: Don't store target type parameters in Config
As the values are only used when a new service is being created, we can
use the Param types to extract the value. Not defining them is much
cleaner than storing them in the configuration as the information is
duplicated elsewhere and the values are used only once.

In the future, the configuration could hold the information that is now
stored elsewhere in the service. For this to be possible, new types for
lists of servers and targets and a monitor type must be added.