MXS-3416: Ignore sescmd results during causal reads
If a causal read was started on a backend while another one was still
executing a session command, the response to the session command would be
misinterpreted as a response to the causal read.

The old code would've worked in 2.4 where the responses to all session
commands were waited on before anything else was executed. In 2.5 as soon
as the expected result is received, the routing will continue. In 2.6
(current develop) the code can be changed to use the result type queue of
the Backend class to know what to do with each result.