MXS-3401: Check unmodifiable service parameters
The target type parameters cannot be modified at runtime and should not be
accepted as valid parameters. Since the core configuration processing
doesn't fully understand parameters that are defined but not used by a
mxs::config::Configuration, the Service itself must check whether they are
defined.

This commit fixes the alter_router test.