MXS-3419: Check permissions with token authentication
The permission checks were missing which caused all users to behave as if
they were allowed to modify MaxScale.

A HTTP 403 error will be returned when token authentication fails due to
missing write access permissions. This is done in order to let the user
know that the authentication itself was successful but the operation
wasn't. This also makes it so that the GUI displays the error instead of
logging the user out.