Add MaskingFilter::post_configure
Doing the work in post_configure() that was previously done in create()
fixes the masking filter test cases. The tests failed because the
parameters in create() are now temporarily empty until they are removed
from the API in favor of doing the configuration via the
mxs::config::Configuration object.

In addition, this makes it easier to enable runtime modification of the
masking filter parameters.