Require suffixes for JSON durations
The REST API always interpreted integer values as milliseconds whereas the
configuration file interpreted it according to the parameter
declaration. This could cause situations where passing the values from the
configuration to MaxCtrl would results in a different, and in most cases
an extremely confusing, result.

Added the change to the upgrading document with a note that the current
support for integer type durations in the configuration file is still
deprecated.

Updated test cases with the required changes.