MXS-3391 Prepare for commands requiring many roundtrips
When tables (aka collections) are auto-created, a command
may have to create a table, but it wont know that before
it has attempted to do whatever it is supposed to do.

Hence, commands need to be able to indicate that they are
busy and that levels above need to wait for additional
responses from the server, before proceeding.