MXS-3332: Fix overlapping capabilities
The RCAP_TYPE_REQUEST_TRACKING capability overlapped with the
RCAP_TYPE_SESSION_STATE_TRACKING capability which enabled the tracking of
OK packet contents for routers that didn't need it, namely readconnroute.

This does not yet explain why the crash happened but it at least should
prevent it from happening in the future.