MXS-3417 Add alternative way to run test
The old TestConnections-constructor does a lot more than a
constructor should as it needs to initialize the whole test system.
It also must call exit() on any errors, as a ctor cannot return
a value.

Now, the recommended way to run a test is to call run_test(). It
performs the initialization, the test and cleanup. It returns the
value from any step that fails.