MXS-2838: Handle hints in binary prepared statements
As the prepared statement ID is assigned before the query is actually
routed, we can preemptively store any hints they contain in the map. If
the prepared statement fails, the hint is erased from the map using the ID
stored in the session. This method optimizes for the case where prepared
statement execution is successful. It comes at the cost of having to
insert and delete a hint that will never be used whenever a prepared
statement fails.