MXS-3427: Make galeramon SQL simpler
The storage of variables and session status in INFORMATION_SCHEMA appears
to be deprecated in MySQL 5.7 and removed in MySQL 8.0 which causes
galeramon to not work with XtraDB Cluster 5.7.

In addition of fixing the support for other databases, simplifying the SQL
to use SHOW VARIABLES and SHOW STATUS also makes it easier to understand
and maintain.