MXS-3435 Prepare for command lookup change
Instead of iterating over the existing commands and then looking
whether a document might match, we will iterate over the keys
of a document and see if we find a corresponding command. This
will be faster as the client libraries appear to put the field
identifying the command first. Further, this will also allow us
to change two lookups into one.