MXS-3436: Fix COM_CHANGE_USER sequence numbers
The backend protocol still assumed that all COM_CHANGE_USER requests
trigger an AuthSwitchRequest which was changed for 2.5 to skip it if it's
the same plugin. This means that the sequence number has to be kept up to
date during the COM_CHANGE_USER which wasn't reset for the initial
authentication exchange.