MXS-3449: Disallow fractional values for seconds
As all duration values are stored in milliseconds in persisted
configurations, disallowing all millisecond values cannot easily be
done. Disallowing all fractional cases that silently round down to zero
solves the main problem while still allowing a warning to be logged for
values which are larger than a second but have a fractional part.