MXS-3435 Add exception classes
Exceptions are already used for handling the case that the type
of an element is not what it should be (implicitly handled by
bsoncxx that throws if you attempt to access a value as the
wrong type).

These classes will be used e.g. for reporting that the query
object is incorrectly built.