MXS-3168: Remove callback from doRequest
The value can be returned as a promise which allows code to be converted
to use the more modern async/await pattern.

Currently the OK response is generated at the lowest level and it should
be moved into the code that displays it to the user. This would mean that
an empty response (i.e. 200 OK) would return a null value inside a
resolved Promise.