Replace Request with Axios
Added backwards compatible functions for interfacing with Axios. Modified
the tests to work correctly with Axios so that all unit tests pass.

Removed the test case with both a Bearer token and basic authentication
credentials. This is not actually possible as both are stored in the
Authorization header and they are mutually exclusive. The fact that it
passed was a side-effect of how Request worked.