MXS-3435 Handle argument documents passed separately
In the case of delete, update, insert the relevant documents
can either be passed in the body of the actual command document,
or separately in the package itself. It seems that the mongodb
client library uses the former approach while the Mongo shell
uses the latter.

Also make some type and presence checking here and there. There
is commonality between the commands, so the common functionality
will be factored out.