MXS-3461: Allow session to be closed on error
Allowing the session to be closed when one of the backend connections
fails makes it possible to use the mirror router to guarantee that each
command is executed on all of the backends. This can be valuable when the
mirror router is used to verify that a migration works as expected.