MXS-3462: Propagate target updates upwards
Whenever a service is updated, all other services that use it must also
recalculate the servers they reach and the capabilities they require.

Adding explicit links to the parent nodes into the service is the simplest
way to propagate the update upwards without having to calculate the graph
at runtime. The only extra work that has to be done is to manage the
parent-child relationship whenever a service is removed or added to a
service.