MXS-3444 Check and set checksum
If the client claims to have provided the checksum, we check it
and if it doesn't match, the connection will be closed. Further,
if the client provided the checksum, we will add it to the response
packet.

Basically I think the server (i.e. mongodbclient) could add the
checksum at its discretion, but testing revealed that at least the
MongoDB nodejs client library (that does not add the checksum) can't
handle responses with a checksum.

Some template trickery had to be added, because now the true type
of the received packed must be passed all the way to the Command
constructor, so that it can be found out whether the response should
include the checksum or not. Will probably be altered later.