MXS-3472: Start trx replay on bad current master
If the current master loses the Master status and no valid replacements
are available at that time, transaction replay should be started. This
fixes the problem where a write query wouldn't be retried if it arrived at
a point in time where the master wasn't valid but an open connection to it
was found.