MXS-3472: Clear server status upon deactivation
As the deactivation is as close to server deletion as we can get, it's
better if a deactivated server is no longer used.

An alternative to simply clearing out the state bits would be to deliver a
hangup event for all connected DCBs but this is a somewhat softer
approach.