MXS-2588: Add Table class
The Table class represents a data stream into an individual table. The
class is the buffer that stores the values until they are transformed into
a suitable form for insertion into MariaDB. It also handles the execution
of all the SQL.

The table creation SQL is compatible with MariaDB 10.2 which also happens
to be the first release with support for bulk execution via
COM_STMT_BULK_EXECUTE. In addition to being the fastest way to insert
data, using the binary protocol also removes the need to escape the JSON
and we can avoid all processing of the data in MaxScale.