MXS-2588: Handle more MariaDB errors
Depending on what part of the JSON is broken, a different error is
returned. We can lump all of the errors under the same umbrella term of
"bad JSON".

Added error handling for failed constraint checks. This happens if the _id
field of the document being inserted is not unique.