MXS-3488 Add configuration option 'insert_behavior'
Specifies whether 'ordered: true' should be handled in a manner
similar to how MongoDB handles it, or in a manner which is efficient
from MariaDB's point of view. That is, whether N documents should
be inserted using N INSERT statements containing just one row of data,
or using 1 INSERT statement containing N rows of data.