MXS-3389 Change GWBUF to mxs::Buffer
As the response buffer is now passed around as a mxs::Buffer, it
will automatically be deleted, yet allow command to take ownership
of the buffer, in case there is a need to.

In the case of cursor handling, there will be a need for keeping
the response buffer around.