MXS-3495: Use a native Kafka installation
Replacing the docker container with a native Kafka installation seems to
be simpler as it removes the problem of "docker" being different on
different platforms. For example, CentOS 8 provides a docker frontend to
podman which behaves differently than the actual docker. This might even
be slightly faster compared to docker which downloaded complete images
whereas the tarball for the kafka binary is less than 70MB.

Also added a flush method to the producer. This is needed to make sure
that all the records are sent before the test starts verifying the
results.

The number of returned records in mxs1687_kafka_ha seems to change from
time to time. This means that there's probably a race condition somewhere
that decides whether the schema is sent or not. With this commit, the test
seems to become relatively unstable.