MXS-3495: Try to consume from Kafka multiple times
There have been cases where the first call to consume_messages() read no
messages but the second call read the expected messages plus the ones from
the first call.

It is possible that the consumption of messages fails due to a timeout
which is expected if Kafka doesn't send or process the message in
time. Repeating the consumption multiple times should help stabilize the
test cases.