MXS-2588: Handle topics that don't exist
Currently the C++ API doesn't have a way to explicitly create topics even
thought it was added to the Kafka API in 0.10.1. This means that if
auto-creation isn't enabled, we can't continue consuming messages even if
some of the topics exist. Enabling auto-creation of topics should help
avoid these situations but it doesn't prevent it from happening. Most
often this happens when the topic is created after MaxScale has been
started.

If topic auto-creation is not enabled, the Kafka consumer won't consume
messages from any of the topics even if some of them exist. To help convey
this to the user, a warning is logged with the topic name that doesn't
exist.

To reduce the amount of logged messages, the consumer will wait the
configured timeout before trying to consume again.

Configured the metadata refreshing to happen more often. This should help
detect created topics faster.