MXS-3495: Stabilize KafkaCDC tests
The test for some reason became very unstable after the switch to the
native Kafka installation. Delaying the starting of MaxScale to until the
replication has been stopped on the slaves and accepting errors about
topics that don't exist seems to improve the situation.

In addition, resetting the replication in mxs1687_kafka_ha seems to solve
the problems with it.