Fix LOAD DATA LOCAL INFILE regression
The backend MariaDB protocol would hit a debug assertion when a LOAD DATA
LOCAL INFILE statement was executed with binary contents. The data was
mistaken for an unknown prepared statement which in turn hit a debug
assertion later on. This fixes the failure of the mxs3459_binary_ldli
test.

This also fixes the double addition of a buffer to the DCB's write
queue. If a query spanned multiple packets, the code didn't correctly
break out after appending the trailing part of the packet.