Split COM_CHANGE_USER handling into two parts
As the COM_CHANGE_USER processing is done using a separate state, the
backend protocol must wait for all pending commands to complete before
generating the COM_CHANGE_USER packet and send it.

Most often this happens when a backend is executing a session command that
has already been executed on all the other backens. This is what happened
in the mxs2295_change_user_loop test which caused a debug assertion to be
triggered.