MXS-3509: Fix syslog logging
The log is now correctly filtered according to the enabled log level. In
addition, the pre-checks for the logging now work due to the correct
handling of session_trace. This means INFO level messages are not
evaluated when log_info is not enabled which could explain why 2.5 could
be slower than 2.4.