MXS-3518: Suspend clients for delayed replies
There are two ways in which libmicrohttpd can handle operations that block
the daemon thread. The first solution is to use the thread-per-connection
mode: each request causes a separate thread to be created and the normal
request handling is executed there. The code can block the thread for as
long as needed and everything works as expected. The downside of this mode
is that it is more wasteful as each request creates a separate thread. In
addition, this mode doesn't support epoll.

The second option is to suspend the connection and perform the blocking
operation in a separate thread. This uses less resources and appears to be
more responsive, at least locally. Since MaxScale already has a thread
pool for various tasks, the response generation can be offloaded
there. This allows the REST API to continue processing new requests.