MXS-3518: Move JWT creation out from Client
The JWT generation can be used elsewhere for the creation of reusable
tokens with some information. In particular, this will be of use when the
right connection must be mapped to the correct client.

Also made the safe usage of these tokens easier by moving the token
splitting code into HttpResponse.