MXS-3443 Make mxg_log_should_log() faster
Heavily called functions make a difference. It would be
even better to move it to the header and inline.

It seems gcc (9.3.0) makes branches where two booleans are
ored or anded together. It probably does so because it could
link to older code that does not obey the current C++ spec.