MXS-3524: Use mxs::Config in log alteration
Using the same mechanism for all alterations of the global configuration
fixes the problem where the persisted value and the runtime value were not
kept in sync. As the log alteration endpoint has used names for the
parameters that weren't the actual parameter names, the code must do some
transformations on the JSON to keep it backwards compatible.

Also fixed the lack of error messages for invalid parameter values for the
log alteration.