MXS-3518: Separate HTTP code from SQL code
Separated the code that deal with the HTTP response generation from the
code that manages the connections. This will help integrate the query
manager implementation into the current code without having to move things
around.

The implementation for testing the query execution is no longer used in
this commit. This means that the query endpoint returns an empty response
after validating the input.