MXS-3518: Split query execution into two parts
Splitting the query endpoint into two parts makes result paginattion
possible and makes the implementation of unique query endpoints
easier. Currently the same endpoint is used for POSTing the query and
GETting the result which isn't ideal but it allows the functionality to be
tested.

When a query pos POSTed, the Location header in the response contains the
URL where the result can be read from. This is a common method of letting
the user know where the resource was created when a user provided
identifier is not used.

The code passes the result pagination information to the lower level code
that should handle the splitting of resultsets into discrete parts. As the
resultset is only navigated once when it is read, only the next link is
created for it.