MXS-3518: Pass connection ID token in request options
As the results are read using a GET request, there is no request body
where the connection ID could be stored. The simplest option is to pass it
along as a parameter in the query. Another option would be to only support
the cookie method of passing information but that makes it harder to use
from non-browser contexts.