MXS-3518: Fix URL creation with implicit API version
If the client used the implicit API version (i.e. didn't pass it as the
first path component), the response would always explicitly include it all
response links. This would cause cookies to not be sent for any links that
were used from a result that was created by an endpoint that used the
implicit version.

For example the /servers/server1/query endpoint would store a cookie but
set the Location header base path as /v1/servers/server1/query which is
not a subpath of the original request. Some cookie jar implementations
only allow subpaths of the originating endpoint to use the cookies. One of
these is the tough-cookie NodeJS library which fails to work when using
the implicit API version but works when it is explicitly defined.

Also fixed the get_uri() method to return a version of the original URI
without the version prefix or leading or trailing slashes. This allows the
host() and get_uri() methods to be used to generate a valid self-link of
any request.