MXS-3518: Return result information to upper layer
The lower level that deals with the concrete results can now passes
information to the upper layer via the boolean pointer. This is needed in
order to know whether the result returned from read_result() was the
complete result or if more data exists.

The concrete implementation will have to either pre-fetch a row to know
whether more rows exist or to count the total number of rows to handle the
case where the row count is the same as the row limit. The former is the
better solution as it allows only a part of the resultset to be read into
memory.