MXS-3518: Make query endpoint JSON API conformant
The JSON API states that a resource that is created without a user given
identifier should return a HTTP 201 Created response with the created
resource in the response body. In addition, the JSON API specification
recommends that the resource collections are in plural form
(i.e. "queries" instead of "query").

An additional benefit of returning the initial response in the POST body
is the elimination of the GET request for small results. This keeps the
API usage relatively simple while still allowing more complex behavior to
be implemented with the pagination links.

One downside of the current implementation is that the results are
readable only once but with the addition of the query resource collection,
it would be possible to keep the results around until the user decides to
delete them.