MXS-3467 Remove `get-panes-dim` event in split-pane
The `get-panes-dim` event approach is not suitable for nested
pane. In this case, the nested panel needs to get its dimension value
but `get-panes-dim` doesn't re-calculate the dimension when the parent
pane dimension changes.

This uses only ref to calculate the dimension in the next tick after
parent panes are rendered or changed it dimensions. The reason why `ref` wasn't
used in the first place because the css `transition` time has an effect
on the accuration of calculating dimension via `ref`.

The virtual-scroll-table component uses header ref to set width of each cell
based on table header's width. So when the pane width is changed ( user resizes pane),
the function to set width of cells needs to be called since `ref` is not
reactive.