MXS-3617 Fix query-result tab items transition
`transition` is disabled as `v-tab-item` uses
`display:none` to not render inactive tab which causes
`setCellWidthMap` function in `virtual-scroll-table`
to set wrong value. This is a limitation of Vue that there
is no way to check DOM elements are completely rendered.
So, when using ref to access DOM and retrieve `DOMRect` (width, height)
when component is mounted which was set `display:none`, the
returns width and height is 0 which is inaccurate.

This disables `transition` in `v-tab-item` and uses `visibility` property
that works with `DOMRect`. This also adds `slide-in` animation to
replace transition animation.