MXS-3518: Return connection resource on connect
As we return a HTTP 201 Created response, the newly created resource must
be present in the response body. To make accessing it easier, the link to
it is also provided in the Location header, as is normally done.

Also added a Location header for the query endpoint to keep it consistent
as we're returning a newly created resource there as well.