MXS-3518: Move query API into its own collection
Treating the database connections as resources makes their use similar to
that of other resources in MaxScale.

The GET operations on connections makes it possible to see how many open
database connections there currently are. Currently the information is not
very useful as it only contains the connection ID. At least the username,
the connection ID and the target where it is connected need to be added to
make it even slightly useful.

In addition, the current connection implementation will remove active
connections from the manager to make sure they cannot be used
concurrently. This causes all active connections to disappear from the GET
output which is misleading if the connection is doing an operation that
will take a long time. The final connection implementation should take
this into account and keep references to "leased" connections.