MXS-3536: Fix max_slave_connections=0
The code always assumed that at least one slave would be required at the
start of a session. This is of course not necessary as readwritesplit can
function with just the master server.

Also removed the unnecessary code from the pre-creation check that
essentially compared that the number of configured servers is not
zero. The connection selection code is able to detect if no valid servers
are available and report the error if needed so the only thing we should
do is check if no servers have been configured.