MXS-3521 Add `resetState` method
This method is used for state hydration. Vuex states are
stored in memory, so when user logouts, all states including
module states need to be returned to initialState.

This also uses localStorage for storing `curr_cnct_resource` state
since the `conn_id_body` token is not a session cookie so `curr_cnct_resource`
needs to be stored until user logout.