MXS-3499: Use COM_QUERY to sync COM_STMT_EXECUTE
By sending a COM_QUERY that closes the connection if the MASTER_GTID_WAIT
times out, we both prevent the execution of the COM_STMT_EXECUTE as well
as trigger the replaying of the query. This has the benefit of leveraging
existing mechanisms while still providing good performance in the case
where the MASTER_GTID_WAIT doesn't time out.