Fully track LOAD DATA LOCAL INFILE
Since the start of the LOAD DATA LOCAL INFILE is detected, the only extra
thing that has to be done is to detect when it ends. By adding the extra
states, we also prevent other problems that might've occurred with the old
implementation that prematurely changed the result state to DONE.

This commit changes the number of "responses" a LOAD DATA LOCAL INFILE
creates from two to one. This is the correct way to represent it as it is
a single command that was split into two for convenience's sake. With the
addition of proper result state tracking, the need for it was removed.

This fixes the rwsplit_infile_outfile test that was broken.