MXS-3541 Extract id from document (field _id)
The id column is now a generated one, which ensures that the
value in the 'id' column is identical with the value of the
'_id' field of the document in the doc column.

As a result of the change, some modifications were required
here and there. Some minor unrelated fixes were made at the
same time. A test file that had not been added, was added
as well.