Remove large packet check from track_query
The state of m_large_query is for the next packet, not the current
packet. Since the trailing parts of large packets aren't passed to
track_query, the check is pointless and only caused problems.

This commit fixes different_size_rwsplit and a handful of other tests that
tested routing of large packets.