Fix debug assert on LOAD DATA LOCAL INFILE
The debug assertion would inspect data that wasn't the first packet of
query which caused it to fail when a LOAD DATA LOCAL INFILE with binary
data was executed.

Also fixed the info logging of the LOAD DATA state: the ending message was
logged in addition to the normal progress message instead of replacing it
when the load was complete.