Prevent double handleError calls
If the result comparison for a command stored in the history fails, the
connection will be closed. This failure wasn't correctly handled and
caused the callback to be installed even in the case where the command
failed. This could cause the callback to close an already closed
connection twice if the response for it arrived before the accepted
one.