Stop processing delayed packets on PS failure
If a prepared statement preparation failed due to a result mismatch, the
state would still go into State::ROUTING and end up handling a
COM_STMT_CLOSE whose ID we do not know. To prevent processing of any
queued packets when a failure occurs, the state should always change into
State::FAILED when handleError is called for the upstream Component.

Added a debug assertion that makes sure the DCB is always closed after the
handleError call completes. This should prevent leakage of connections
that can happen if the router doesn't properly close the downstream
Endpoint.