Adding more features to editor and modes to run
When `Place Schema in Editor` is called, this inserts schema text
at current cursor position instead of concatenating string.

This add two modes to dispatch `fetchQueryResult` which are
`all` and `selected`. The former dispatches query with all statements
while the latter dispatches with selected statements in the editor.

This also adds custom actions(shortkeys) with the following keybindings to execute
query:
- Ctrl/CMD + Enter: Execute all statements
- Ctrl/CMD + Shift + Enter: Execute selected statements.