Adjust debug assertion in do_handle_error
It turns out that routers that unconditionally return false from
handleError (i.e. they don't handle errors) do not close the DCB which
triggered the debug assertion. Since technically this is correct behavior
for the router, the debug assertion must be adjusted.

Also added the logging of the current session ID, the service it connects
to and the DCB pointer to the fatal signal handler. As this information is
often available, it's helpful if it's logged when there's more than one
service in MaxScale.