Create volatile nodes during configuration
This isn't ideal but it removes the need to communicate with the
MainWorker.

Now that the creation of the volatile servers is done in the MainWorker
thread, runtime_create_volatile_server can assert that it is only called
from the MainWorker thread and nowhere else. This also means that the
run_in_mainworker function must always queue the execution of the function
to make sure the configuration processing has finished before the server
is added to all services.