Fix mxs2729_rebalance
The test mistakenly gave the rebalance_threshold as seconds instead of the
integer it was supposed to be. This wasn't detected by the test as it was
using the plain maxctrl function instead of the check_maxctrl one.

Also changed the value of rebalance_window to be twice that of
rebalance_period to keep the window larger than the period as is suggested
by the documentation.