Fix pam_authentication
The test used both the normal constructor with argc and argv as well as
the run_test function. This caused initialization to be done twice which
eventually caused the test raise an exception when a joinable thread was
assigned a new value.

Added safeguards that help detect this in a more obvious way.