MXS-3467 Fix frozen issue in result-data-table
When resizing pane fastly, `virtual-scroll-table` is crashed and
frozen due to negative height of the table. Though, this only
occurs when the table is handling more than 10000 rows of data.
Even the table is using virtual scroll approach, the result
data is kept in memory which causes issue to `virtual-scroll-table`.

Ideally, pagination should be used. i.e. Keep the number of items in
array to not exceed 10000 rows in memory (perhaps, it should be measured
in unit of data .e.g. 1MB).
When scrolling position reaches the bottom or top, fetching the next page
or prev page then replacing current result data with new chunk of data.

This also uses JSON.parse then JSON.stringify to deep clone table rows
because lodash cloneDeep is slower and it causes laggy issue when
dealing with large data.