Escape identifier of db_tree
This creates `escapeIdentifiers` helper function to escape
id of each node in db_tree when it is used as an indentifier.
This also escapes `insertText` value for autocompletion list
of editor. `label` will be shown as unescaped text but when
it is inserted to the editor, the value will be escaped.