Redesign `connection-manager` UI
Since we only allow one connection at a time, displaying it as
a dropdown is not suitable. So, the connection management is
redesigned with button. i.e  Alloing to open new connection
when current one is disconnected or when there is no
active connection.

When there is active connection, it shows the name
of the resource being connected and the `unlink` icon to
disconnect.

The button will show `Open connection` after user successfully
disconnect the current active connection. Clicking this will
open `connection-dialog`.