Call post_configure() only when a change occurred
If the call to configure() doesn't cause a parameter to change its value,
don't call it. This check doesn't need to be done for the ConfigParameters
version as it's guaranteed to be called only once.

This caused a debug assertion to be hit when an attempt to alter the
qlafilter was made with the same parameters it was already using.