MXS-2808: Store current configuration as JSON
The REST API stores the latest configuration as JSON whenever it receives
a request that modifies data. No-op operations aren't currently detected
and cause the configuration to be saved.

The configuration is formed using the existing REST API code which removes
the need to write similar code twice. The downside of this is that we
don't want to keep all of the information stored in the responses and, to
minimize the amount of data stored, the results have to go through a
pruning step.

If a dynamically generated configuration file exists, MaxScale will
attempt to load it instead of the normal configuration file. This prevents
the possibility of conflicts arising from the difference between the
static and the dynamic configurations. The configuration loading code is
not yet implemented.