Move code into maxbase/json.hh
Moved the JSON dumping code into maxbase as it's generic enough to be
there. Also switched the dumping function to the callback version which
allows the result to be constructed in place.

Added the mxb::Json::to_string() method as a convenient wrapper for it.

Added mxb::Json::get_json(). It returns the underlying pointer to the
json_t object and it allows mxb::Json to be used to navigate to a known
position in the object. This can then be passed elsewhere in non-mxb::Json
form.