Fix reference leak in mxb::Json::load
The function leaked a reference when the json_null() that reset()
allocated was unconditionally overwritten in load(). This didn't cause a
memory leak as the value returned by json_null() is a pointer to a static
struct.

To prevent this and to also retain the current state, the reset() can be
modified to behave similarly to how the std::unique_ptr::reset method
works: free the current object and replace it with the new one. This
allows an attempt to load JSON from a string or a file to be made without
loosing the current object.