Extend mxb::Json
Created const lvalue overloads for versions that used a rvalue as their
input. This makes their use easier when both the source and the
destination must be kept, for example when one Json is copied to multiple
Jsons.

Adding the set_float method allows all valid JSON types to be
set. Currently not used but will be useful when the statistics JSON is
generated using this class.

The erase method is needed when objects that were loaded must be modified
before being used. This will be of particular use when loading the
configuration from JSON.

The Type::NONE removes the need to allocate an object before loading the
contents from a string or a file. Not absolutely necessary but it's
clearer to have an explicit empty state rather than an implicit one.

Also documented all methods.