MXS-2808: Move config creation into ConfigManager
The objects are now stored in a flat array instead of being partitioned by
their type. This allows the objects to be stored in the order in which
they must be constructed which is easier to resolve during the
configuration creation than it is to resolve them from the JSON.

Currently the objects are stored in a static order. This will cause
service creation to fail if a service uses a parameter that refers to
another service and the services aren't ordered correctly. Currently this
isn't a problem as none of the services have such parameters and they only
refer to each other via the relationships.

Having an object that manages the creation and loading of the
configuration from JSON makes it simpler to retain the latest JSON
configuration. Currently it isn't stored but the possibility is there.