MXS-2808: Add skeleton config processing code
The code figures out which objects must be removed and which ones are
new. The order in which the deletion is done is not super important if the
config_runtime.hh functions are used as they are capable of severing
linkages between objects if called with the force flag set to true.