MXS-2808: Add removal of old objects
The old object removal is the first step in the configuration update
process. This must be done before new objects are created in case a
"unique" resource is claimed by an object. Current only servers and
listeners are affected: the address and port uniquely identifies both
servers and listeners.

Added a simple exception class for dealing with errors encountered during
the configuration processing.