MXS-2808: Add creation of new objects
Most of the objects can be created with the relationships still in
place. The exception to this rule are services, which must not have any
relationships, and monitors, which must not have service relationships but
should have the server relationships. Listener require the service
relationship to exists when they are created which is why they are the
last objects to be created.

As the creation of services is done in two steps, the current code will
not result in a functional configuration if it were enabled. The update
step is required to form the links between the services and the other
objects they use.