MXS-2808: Move code into ConfigManager
Moved the code into ConfigManager and made it a non-singleton
class. Having the required components be arguments for the constructor
makes sure the initialization order is correct.

Renamed the functions to better reflect what they are used for. The extra
entry points will eventually be needed but are mostly empty for the time
being.