MXS-2808: Add exception throwing helper
Simplifies its usage by moving the error formatting into a variadic
function. Having the actual throw in the function would be convenient but
isn't as clear as throwing the return value of the function. This makes it
very clear that an exception is being thrown.

Also removed the auto return type from remove_extra_data, it was not
intentional.