MXS-2808: Add SQL creation functions
The UPDATE SQL is the same as the one that was originally planned. To
handle the case where there are no rows, a separate INSERT statement is
simpler than using any of the UPSERT variants available in MariaDB.

The synchronization is done using two SELECT statements. This makes it
possible to detect when the configuration in the cluster is older than the
local configuration in which case a warning can be logged. If this is
deemed not important, the synchronization could be done with only the
SELECT statement that returns the new configuration if one is available.