MXS-2808: Revert bad configuration updates
If a configuration in the cluster cannot be applied, an attempt to revert
to the old configuration is made. This allows the local MaxScale to settle
to a known good configuration version.

It is still possible for MaxScale to end up in a state that's between the
old and the new configuration if the reverting process also fails. Both in
this case and in the successful revert case each configuration update is
only attempted once. This will help avoid repeated attempts to apply the
same configuration which could make the situation worse.