MXS-2808: Add ConfigManager diagnostic output
This can be used to check if a MaxScale instance has diverged from the
cluster. Using a SHA1 checksum is preferable to a CRC32 checksum as one
can calculate the checksum of the config stored in the database with:

  SELECT SHA1(config) FROM mysql.maxscale_config;

If the checksum is not the same, it doesn't always mean that the
configuration has diverged: any global parameters that cannot be modified
at runtime (e.g. datadir) that aren't the exact same value will cause the
checksum to be different from the one stored in the cluster. This could be
avoided by not including the static parameters in the configuration.